Update coordinates of JavaMail dependency
JavaMail is moving towards using javax.mail:javax.mail-api as the
coordinates for the JavaMail API (1.5 no longer publishes the
javax.mail:mail artifact). This commit updates spring-cloud to use
the new coordinates but leaves the version unchanged.

The dependency has also been moved into the testRuntime configuration
as it is not needed for the tests to compile.