Add the Spring IO plugin
This commit adds the Spring IO plugin, but it is configured such that
it will only take effect when the build is run with
-PplatformVersion=x.y.z

Some redundant exclusions in spring-cloud-spring-service-connector have
been removed and those that only declared a module have been updated to
declare both a module and a group.

The Spring IO Platform currently specifies a newer version (2.12.1) of
the MongoDB driver. This version has changed the behaviour of
Mongo.getAddress() such that calling it now results in an attempt to
initiate the connection. MongoServiceConnectorCreatorTest has been
updated to call Mongo.getAllAddresses() instead. This works on both
2.11.x and 2.12.1.