Drop spring-amqp dependency version down to 1.1.1
Spring Cloud Connector dependencies on other Spring projects (e.g. Spring AMQP, Spring Data) are not transient, but only necessary for compiling the Spring Cloud Connector code itself. Users of Spring Cloud Connectors are responsible for providing their own Spring dependencies of an appropriate version. Spring Cloud Connector should depend on the oldest version of a Spring project that provides all the APIs necessary for compilation. In the case of Spring AMQP, this is 1.1.1.