Improve Key-Value Backwards Compatibility
Currently, the entire Spring Service Connector implementation is backwards
compatible to Spring Data 1 except for the Key Value (Redis) services.  This
is broken because the Spring Data 2 Redis API has been improved.  It turns out
that the Spring 2 Redis API doesn't remove the previous Spring Data 1 variant
and therefore, the two can coexist.

This change reinstates the previous Spring Data 1 Redis implementation and
adds it to the list of ServiceConnectorCreator services.  If the classpath
contains the Spring Data 2 variant, then only the Spring Data 2 Creator will
be used.  If the classpath contains the Spring Data 1 variant, then only the
Spring Data 1 Creator will be used.