Shrink Cloud Foundry Connector Footprint
Previously, the Cloud Foundry Connector utilized Jackson for parsing the
VCAP_SERVICES JSON payload.  While Jackson should be the default choice for
most JSON parsing, the fact that the project does not need Object binding
means that it is overkill for the need.  In addition, since whichever JSON
parser is chosen must be included in the distributable JAR and not shared with
the rest of the application (which is likely to have its own version of
Jackson) the size of the Jackson libraries becomes a significant concern.

This change updates the code to use the org.json:json parser and shrinks the
spring-cloud-cloudfoundry-connector from 1.2M to 84K with no difference in
functionality.

[resolves #73]